refactor: simplify build and packaging of navbar
