+++ packages/components/navbar/bs-navbar-link.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsNavbarLinkCss } from './bs-navbar-link.css.js';
+import { BsNavbarLinkCss } from './css/bs-navbar-link.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsNavbarLink extends LitElement {
