+++ packages/components/navbar/bs-navbar-nav.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsNavbarNavCss } from './bs-navbar-nav.css.js';
+import { BsNavbarNavCss } from './css/bs-navbar-nav.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsNavbarNav extends LitElement {
