+++ packages/components/navbar/bs-navbar-text.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsNavbarTextCss } from './bs-navbar-text.css.js';
+import { BsNavbarTextCss } from './css/bs-navbar-text.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsNavbarText extends LitElement {
