+++ packages/components/navbar/css/index.js
@@ -1,9 +1,6 @@
 export { BsNavbarBrandLinkCss } from './bs-navbar-brand-link.css.js';
-export { BsNavbarBrandLink } from './bs-navbar-brand-link.js';
 export { BsNavbarBrandCss } from './bs-navbar-brand.css.js';
-export { BsNavbarBrand } from './bs-navbar-brand.js';
 export { BsNavbarCollapseCss } from './bs-navbar-collapse.css.js';
-export { BsNavbarCollapse } from './bs-navbar-collapse.js';
 export { BsNavbarDarkColorCss } from './bs-navbar-dark-color.css.js';
 export { BsNavbarExpandLargeCss } from './bs-navbar-expand-lg.css.js';
 export { BsNavbarExpandMediumCss } from './bs-navbar-expand-md.css.js';
@@ -11,14 +8,8 @@ export { BsNavbarExpandSmallCss } from './bs-navbar-expand-sm.css.js';
 export { BsNavbarExpandExtraSmallCss } from './bs-navbar-expand-xs.css.js';
 export { BsNavbarLightColorCss } from './bs-navbar-light-color.css.js';
 export { BsNavbarLinkCss } from './bs-navbar-link.css.js';
-export { BsNavbarLink } from './bs-navbar-link.js';
 export { BsNavbarNavCss } from './bs-navbar-nav.css.js';
-export { BsNavbarNav } from './bs-navbar-nav.js';
 export { BsNavbarTextCss } from './bs-navbar-text.css.js';
-export { BsNavbarText } from './bs-navbar-text.js';
 export { BsNavbarTogglerIconCss } from './bs-navbar-toggler-icon.css.js';
-export { BsNavbarTogglerIcon } from './bs-navbar-toggler-icon.js';
 export { BsNavbarTogglerCss } from './bs-navbar-toggler.css.js';
-export { BsNavbarToggler } from './bs-navbar-toggler.js';
 export { BsNavbarCss } from './bs-navbar.css.js';
-export { BsNavbar } from './bs-navbar.js';
